+++ lib/plausible/billing/qouta/usage.ex
@@ -1,34 +1,26 @@
-defmodule Plausible.Billing.Quota do
-  @moduledoc """
-  This module provides functions to work with plans usage and limits.
-  """
+defmodule Plausible.Billing.Quota.Usage do
+  @moduledoc false
 
   use Plausible
   import Ecto.Query
   alias Plausible.Users
   alias Plausible.Auth.User
   alias Plausible.Site
-  alias Plausible.Billing.{Plan, Plans, Subscription, Subscriptions, EnterprisePlan, Feature}
-  alias Plausible.Billing.Feature.{Goals, RevenueGoals, Funnels, Props, StatsAPI}
+  alias Plausible.Billing.{Subscriptions}
+  alias Plausible.Billing.Feature.{RevenueGoals, Funnels, Props, StatsAPI}
 
-  @type limit() :: :site_limit | :pageview_limit | :team_member_limit
+  @type cycles_usage() :: %{cycle() => usage_cycle()}
 
-  @type over_limits_error() :: {:over_plan_limits, [limit()]}
+  @typep cycle :: :current_cycle | :last_cycle | :penultimate_cycle
+  @typep last_30_days_usage() :: %{:last_30_days => usage_cycle()}
+  @typep monthly_pageview_usage() :: cycles_usage() | last_30_days_usage()
 
-  @type monthly_pageview_usage() :: %{period() => usage_cycle()}
-
-  @type period :: :last_30_days | :current_cycle | :last_cycle | :penultimate_cycle
-
-  @type usage_cycle :: %{
-          date_range: Date.Range.t(),
-          pageviews: non_neg_integer(),
-          custom_events: non_neg_integer(),
-          total: non_neg_integer()
-        }
-
-  @pageview_allowance_margin 0.1
-
-  def pageview_allowance_margin(), do: @pageview_allowance_margin
+  @typep usage_cycle :: %{
+           date_range: Date.Range.t(),
+           pageviews: non_neg_integer(),
+           custom_events: non_neg_integer(),
+           total: non_neg_integer()
+         }
 
   def usage(user, opts \\ []) do
     basic_usage = %{
@@ -45,108 +37,12 @@ defmodule Plausible.Billing.Quota do
     end
   end
 
-  on_ee do
-    @limit_sites_since ~D[2021-05-05]
-    @site_limit_for_trials 10
-    @team_member_limit_for_trials 3
-
-    @spec site_limit(User.t()) :: non_neg_integer() | :unlimited
-    def site_limit(user) do
-      if Timex.before?(user.inserted_at, @limit_sites_since) do
-        :unlimited
-      else
-        get_site_limit_from_plan(user)
-      end
-    end
-
-    defp get_site_limit_from_plan(user) do
-      user = Users.with_subscription(user)
-
-      case Plans.get_subscription_plan(user.subscription) do
-        %{site_limit: site_limit} -> site_limit
-        :free_10k -> 50
-        nil -> @site_limit_for_trials
-      end
-    end
-
-    @spec team_member_limit(User.t()) :: non_neg_integer()
-    def team_member_limit(user) do
-      user = Users.with_subscription(user)
-
-      case Plans.get_subscription_plan(user.subscription) do
-        %{team_member_limit: limit} -> limit
-        :free_10k -> :unlimited
-        nil -> @team_member_limit_for_trials
-      end
-    end
-  else
-    def site_limit(_) do
-      :unlimited
-    end
-
-    def team_member_limit(_) do
-      :unlimited
-    end
-  end
-
   @spec site_usage(User.t()) :: non_neg_integer()
   @doc """
   Returns the number of sites the given user owns.
   """
   def site_usage(user) do
     Plausible.Sites.owned_sites_count(user)
-  end
-
-  @doc """
-  Enterprise plans are always allowed to add more sites (even when
-  over limit) to avoid service disruption. Their usage is checked
-  in a background job instead (see `check_usage.ex`).
-  """
-  def ensure_can_add_new_site(user) do
-    user = Users.with_subscription(user)
-
-    case Plans.get_subscription_plan(user.subscription) do
-      %EnterprisePlan{} ->
-        :ok
-
-      _ ->
-        usage = site_usage(user)
-        limit = site_limit(user)
-
-        if below_limit?(usage, limit), do: :ok, else: {:error, {:over_limit, limit}}
-    end
-  end
-
-  @monthly_pageview_limit_for_free_10k 10_000
-  @monthly_pageview_limit_for_trials :unlimited
-
-  @spec monthly_pageview_limit(User.t() | Subscription.t()) ::
-          non_neg_integer() | :unlimited
-  def monthly_pageview_limit(%User{} = user) do
-    user = Users.with_subscription(user)
-    monthly_pageview_limit(user.subscription)
-  end
-
-  def monthly_pageview_limit(subscription) do
-    case Plans.get_subscription_plan(subscription) do
-      %EnterprisePlan{monthly_pageview_limit: limit} ->
-        limit
-
-      %Plan{monthly_pageview_limit: limit} ->
-        limit
-
-      :free_10k ->
-        @monthly_pageview_limit_for_free_10k
-
-      _any ->
-        if subscription do
-          Sentry.capture_message("Unknown monthly pageview limit for plan",
-            extra: %{paddle_plan_id: subscription.paddle_plan_id}
-          )
-        end
-
-        @monthly_pageview_limit_for_trials
-    end
   end
 
   @doc """
@@ -195,8 +91,7 @@ defmodule Plausible.Billing.Quota do
     end
   end
 
-  @spec usage_cycle(User.t(), period(), list() | nil, Date.t()) :: usage_cycle()
-
+  @spec usage_cycle(User.t(), :last_30_days | cycle(), list() | nil, Date.t()) :: usage_cycle()
   def usage_cycle(user, cycle, owned_site_ids \\ nil, today \\ Timex.today())
 
   def usage_cycle(user, cycle, nil, today) do
@@ -398,114 +293,9 @@ defmodule Plausible.Billing.Quota do
     for {f_mod, used?} <- used_features, used?, f_mod.enabled?(site), do: f_mod
   end
 
-  @doc """
-  Ensures that the given user (or the usage map) is within the limits
-  of the given plan.
-
-  An `opts` argument can be passed with `ignore_pageview_limit: true`
-  which bypasses the pageview limit check and returns `:ok` as long as
-  the other limits are not exceeded.
-  """
-  @spec ensure_within_plan_limits(User.t() | map(), struct() | atom() | nil, Keyword.t()) ::
-          :ok | {:error, over_limits_error()}
-  def ensure_within_plan_limits(user_or_usage, plan, opts \\ [])
-
-  def ensure_within_plan_limits(%User{} = user, %plan_mod{} = plan, opts)
-      when plan_mod in [Plan, EnterprisePlan] do
-    ensure_within_plan_limits(usage(user), plan, opts)
-  end
-
-  def ensure_within_plan_limits(usage, %plan_mod{} = plan, opts)
-      when plan_mod in [Plan, EnterprisePlan] do
-    case exceeded_limits(usage, plan, opts) do
-      [] -> :ok
-      exceeded_limits -> {:error, {:over_plan_limits, exceeded_limits}}
-    end
-  end
-
-  def ensure_within_plan_limits(_, _, _), do: :ok
-
-  defp exceeded_limits(usage, plan, opts) do
-    for {limit, exceeded?} <- [
-          {:team_member_limit, not within_limit?(usage.team_members, plan.team_member_limit)},
-          {:site_limit, not within_limit?(usage.sites, plan.site_limit)},
-          {:monthly_pageview_limit,
-           exceeds_monthly_pageview_limit?(usage.monthly_pageviews, plan, opts)}
-        ],
-        exceeded? do
-      limit
-    end
-  end
-
-  defp exceeds_monthly_pageview_limit?(usage, plan, opts) do
-    if Keyword.get(opts, :ignore_pageview_limit) do
-      false
-    else
-      case usage do
-        %{last_30_days: %{total: total}} ->
-          !within_limit?(total, pageview_limit_with_margin(plan, opts))
-
-        billing_cycles_usage ->
-          Plausible.Workers.CheckUsage.exceeds_last_two_usage_cycles?(
-            billing_cycles_usage,
-            plan.monthly_pageview_limit
-          )
-      end
-    end
-  end
-
-  defp pageview_limit_with_margin(%{monthly_pageview_limit: limit}, opts) do
-    margin = Keyword.get(opts, :pageview_allowance_margin, @pageview_allowance_margin)
-    ceil(limit * (1 + margin))
-  end
-
-  @doc """
-  Returns a list of features the user can use. Trial users have the
-  ability to use all features during their trial.
-  """
-  def allowed_features_for(user) do
-    user = Users.with_subscription(user)
-
-    case Plans.get_subscription_plan(user.subscription) do
-      %EnterprisePlan{features: features} ->
-        features
-
-      %Plan{features: features} ->
-        features
-
-      :free_10k ->
-        [Goals, Props, StatsAPI]
-
-      nil ->
-        if Users.on_trial?(user) do
-          Feature.list()
-        else
-          [Goals]
-        end
-    end
-  end
-
   defp owned_sites_query(user) do
     from sm in Site.Membership,
       where: sm.role == :owner and sm.user_id == ^user.id,
       select: %{site_id: sm.site_id}
-  end
-
-  @spec below_limit?(non_neg_integer(), non_neg_integer() | :unlimited) :: boolean()
-  @doc """
-  Returns whether the usage is below the limit or not.
-  Returns false if usage is equal to the limit.
-  """
-  def below_limit?(usage, limit) do
-    if limit == :unlimited, do: true, else: usage < limit
-  end
-
-  @spec within_limit?(non_neg_integer(), non_neg_integer() | :unlimited) :: boolean()
-  @doc """
-  Returns whether the usage is within the limit or not.
-  Returns true if usage is equal to the limit.
-  """
-  def within_limit?(usage, limit) do
-    if limit == :unlimited, do: true, else: usage <= limit
   end
 end