+++ test/plausible_web/controllers/api/external_stats_controller/auth_test.exs
@@ -150,6 +150,7 @@ defmodule PlausibleWeb.Api.ExternalStatsController.AuthTest do
     })
   end
 
+  @tag :ee_only
   test "returns HTTP 402 when user is on a growth plan", %{
     conn: conn,
     user: user,