+++ test/plausible_web/controllers/site_controller_test.exs
@@ -364,7 +364,7 @@ defmodule PlausibleWeb.SiteControllerTest do
         })
 
       assert redirected_to(conn) == "/example.com/snippet?site_created=true"
-      assert Plausible.Billing.Quota.site_usage(user) == 3
+      assert Plausible.Billing.Quota.Usage.site_usage(user) == 3
     end
 
     for url <- ["https://Example.com/", "HTTPS://EXAMPLE.COM/", "/Example.com/", "//Example.com/"] do
@@ -485,6 +485,31 @@ defmodule PlausibleWeb.SiteControllerTest do
       assert resp =~ "Site Timezone"
       assert resp =~ "Site Domain"
       assert resp =~ "JavaScript Snippet"
+    end
+  end
+
+  describe "GET /:website/settings/people" do
+    setup [:create_user, :log_in, :create_site]
+
+    @tag :ee_only
+    test "shows members page with links to CRM for super admin", %{
+      conn: conn,
+      user: user,
+      site: site
+    } do
+      patch_env(:super_admin_user_ids, [user.id])
+
+      conn = get(conn, "/#{site.domain}/settings/people")
+      resp = html_response(conn, 200)
+
+      assert resp =~ "/crm/auth/user/#{user.id}"
+    end
+
+    test "does not show CRM links to non-super admin user", %{conn: conn, user: user, site: site} do
+      conn = get(conn, "/#{site.domain}/settings/people")
+      resp = html_response(conn, 200)
+
+      refute resp =~ "/crm/auth/user/#{user.id}"
     end
   end